first upload
This is a website that has a small chat window that only holds 10 messages on the server at a time. The messages are stored unecrypted in the messages.txt file. Each session can see what another session types, and every 10th message the oldest message is deleted. Each user is given a random username that resembles a phone number from the 517 area code.